requirements_for_tts.txt
coqui-tts==0.24.2
coqui-tts-trainer==0.1.4